refactor discord feature
