+++ src/main/java/de/blutmondgilde/modpackaddons/discord/utils/StringUtils.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.gitlab.cdagaming.craftpresence.utils;
+package de.blutmondgilde.modpackaddons.discord.utils;
 
 import java.util.regex.Pattern;
